Compiles but no window
